+++ src/components/core/media/FileDetails/FileDetails-types.ts
@@ -1,5 +1,5 @@
 interface ImageFile {
-  id: string;
+  fileId: string;
   created: Date;
   modified: Date;
   type: string;